Refactor commands into the game package
