+++ game/cmd_matchers.go
@@ -1,4 +1,4 @@
-package commands
+package game
